CogVM source as per VMMaker.oscog-eem.992
64-bit Spur:
Correct a few comparisons and shifts in small float
code to make comparisons in generated C correct.

Move loadFloatOrInt: into the ObjectMemory hierarchies to allow
Spur64BitMemoryManager to generate better code.

Make printOopShortInner: small-float savvy.

Make printHexnp: 64-bit aware.

Can now interact with a 64-bit Spur image.

Unix:
Fix missed signature change in sqUnixFBDev.c.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3192 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 184b1013d52fc4be775e5c64f614363557669216

+++ platforms/unix/vm-display-fbdev/sqUnixFBDev.c
@@ -418,7 +418,7 @@ static sqInt display_ioForceDisplayUpdate(void)								{ return 0; }
 static sqInt display_ioSetDisplayMode(sqInt width, sqInt height, sqInt depth, sqInt fullscreenFlag)		{ return 0; }
 static void display_winSetName(char *imageName)								{ return  ; }
 static void display_winExit(void)									{ return  ; }
-static int  display_winImageFind(char *buf, int len)							{ return 0; }
+static long  display_winImageFind(char *buf, int len)							{ return 0; }
 static void display_winImageNotFound(void)								{ return  ; }
